Move data type constants to constants file
