+++ frontend/src/components/eMIB/constants.js
@@ -3,6 +3,16 @@ export const HEADER_HEIGHT = 130;
 export const FOOTER_HEIGHT = 72;
 
 // Data structure
+export const ACTION_TYPE = {
+  email: "email",
+  task: "task"
+};
+
+export const EDIT_MODE = {
+  create: "create",
+  update: "update"
+};
+
 export const EMAIL_TYPE = {
   reply: "reply",
   replyAll: "reply all",
